Fixed bugs on newplan tests
